basic pundit working on turnsheet

+++ app/controllers/turnsheets_controller.rb
@@ -1,5 +1,6 @@
 class TurnsheetsController < ApplicationController
   before_action :set_turnsheet, only: %i[ show edit update destroy ]
+  before_action :authenticate_user!
 
   # GET /turnsheets or /turnsheets.json
   def index
@@ -8,11 +9,13 @@ class TurnsheetsController < ApplicationController
 
   # GET /turnsheets/1 or /turnsheets/1.json
   def show
+    authorize @turnsheet
   end
 
   # GET /turnsheets/new
   def new
     @turnsheet = Turnsheet.new
+    authorize @turnsheet
   end
 
   # GET /turnsheets/1/edit
@@ -22,6 +25,7 @@ class TurnsheetsController < ApplicationController
   # POST /turnsheets or /turnsheets.json
   def create
     @turnsheet = Turnsheet.new(turnsheet_params)
+    authorize @turnsheet
 
     respond_to do |format|
       if @turnsheet.save
@@ -50,6 +54,7 @@ class TurnsheetsController < ApplicationController
   # DELETE /turnsheets/1 or /turnsheets/1.json
   def destroy
     @turnsheet = Turnsheets.find(params[:id])
+    authorize @turnsheet
 
     if @turnsheet.destroy
       # Successful deletion
@@ -63,34 +68,34 @@ class TurnsheetsController < ApplicationController
 
   def process_turnsheet
     errors = [] # Initialize an empty array to store any errors
-  
+
     Turnsheet.find_each do |turnsheet|
       next if turnsheet.processed.present?
-  
+
       turnsheet.save # Save the Turnsheet record first
-  
+
       Selection.where(club: turnsheet.club).destroy_all
-  
+
       (1..11).each do |i|
         Selection.create(club: turnsheet.club, player_id: turnsheet.send("player_#{i}"), turnsheet: turnsheet)
       end
-  
+
       if turnsheet.coach_upg.present?
         Turn.create({ week: turnsheet.week, club: turnsheet.club, var1: 'coach', var2: turnsheet.coach_upg, var3: 500000, turnsheet: turnsheet })
       end
-  
+
       if turnsheet.prop_upg.present?
         Turn.create({ week: turnsheet.week, club: turnsheet.club, var1: 'prop', var2: turnsheet.prop_upg, var3: 250000, turnsheet: turnsheet })
       end
-  
+
       if turnsheet.train_gkp.present?
         Turn.create({ week: turnsheet.week, club: turnsheet.club, var1: 'train', var2: turnsheet.train_gkp, var3: turnsheet.train_gkp_skill, turnsheet: turnsheet })
       end
-  
+
       if turnsheet.train_dfc.present?
         Turn.create({ week: turnsheet.week, club: turnsheet.club, var1: 'train', var2: turnsheet.train_dfc, var3: turnsheet.train_dfc_skill, turnsheet: turnsheet })
       end
-  
+
       if turnsheet.train_mid.present?
         Turn.create({ week: turnsheet.week, club: turnsheet.club, var1: 'train', var2: turnsheet.train_mid, var3: turnsheet.train_mid_skill, turnsheet: turnsheet })
       end
@@ -98,7 +103,7 @@ class TurnsheetsController < ApplicationController
       if turnsheet.train_att.present?
         Turn.create({ week: turnsheet.week, club: turnsheet.club, var1: 'train', var2: turnsheet.train_att, var3: turnsheet.train_att_skill, turnsheet: turnsheet })
       end
-  
+
       if turnsheet.stad_upg.present?
         Turn.create({ week: turnsheet.week, club: turnsheet.club, var1: turnsheet.stad_upg, var2: turnsheet.stad_amt, var3: turnsheet.val, turnsheet: turnsheet })
       end
@@ -106,23 +111,23 @@ class TurnsheetsController < ApplicationController
       if turnsheet.stad_cond_upg.present?
         Turn.create({ week: turnsheet.week, club: turnsheet.club, var1: turnsheet.stad_cond_upg, var3: 100000, turnsheet: turnsheet })
       end
-  
+
       turnsheet.update(processed: DateTime.now)
     rescue StandardError => e
       errors << "Error processing turnsheet with ID #{turnsheet.id}: #{e.message}"
     end
-  
+
     if errors.empty?
       notice = "Turnsheets were processed successfully."
     else
       notice = "Errors occurred while processing turnsheets:\n\n#{errors.join("\n")}"
     end
-  
+
     respond_to do |format|
       format.html { redirect_to turnsheets_url, notice: notice }
     end
   end
-  
+
 
   private
